add if statement to convert multiline to singleline in case input is already single line

+++ workflow/rules/2_derive_variants.smk
@@ -56,15 +56,18 @@ rule gen_derived:
 		"results/derived_variants/raw/chr{chr}.vcf",
 	shell:
 		'''
-		echo "Formatting multiline fasta to single line fasta"
-		# this one needs a failsafe if genome is already a oneliner!!!!!
+		if [ `wc -l file.txt | awk '{print $1}'` -ge "3" ]
+		then
+			echo "Formatting multiline fasta to single line fasta"
+			# this one needs a failsafe if genome is already a oneliner!!!!!
 
-		start=$(date +%s)
-		awk '/^>/ {{printf("\\n%s\\n",$0); next; }} {{ printf("%s",$0);}} END {{printf("\\n");}}' {input.reference} > tmp{wildcards.chr}
-		mv tmp{wildcards.chr} {input.reference}
-		end=$(date +%s)
-		echo "Elapsed time: $(($end-$start)) seconds"
-  
+			start=$(date +%s)
+			awk '/^>/ {{printf("\\n%s\\n",$0); next; }} {{ printf("%s",$0);}} END {{printf("\\n");}}' {input.reference} > tmp{wildcards.chr}
+			mv tmp{wildcards.chr} {input.reference}
+			end=$(date +%s)
+			echo "Elapsed time: $(($end-$start)) seconds"
+  		fi
+
 		python3 {input.script} \
 		 -c {wildcards.chr} \
 		 -a {input.ancestral} \
